split core with directives, fixing typings files
