+++ .gitignore
@@ -12,8 +12,11 @@ pids
 .bin
 bin
 reports
-
-*.compiled.js
+directives
+Decorators.d.ts
+Decorators.js
+Slim.js
+Slim.d.ts
 
 .idea
 .vscode
@@ -29,18 +32,8 @@ src/playground
 # nyc test coverage
 .nyc_output
 
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
 # Dependency directories
 node_modules
-jspm_packages
 
 # Optional npm cache directory
 .npm
