+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2016 Avichay Eyal
+Copyright (c) 2016 Eyal Avichay <eavichay@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
