+++ package.json
@@ -2,13 +2,12 @@
   "name": "slim-js",
   "version": "3.3.4",
   "description": "Superpower Web Components",
-  "main": "Slim.min.js",
+  "main": "Slim.js",
   "scripts": {
-    "dev": "babel src/Slim.js --watch --out-file Slim.min.js",
-    "build:compile": "babel src/Slim.js > Slim.js",
-    "build:uglify": "uglifyjs Slim.js -o Slim.min.js",
-    "build:es6": "uglifyjs src/Slim.js -o Slim.es6.js",
-    "build": "npm run build:compile && npm run build:uglify && npm run build:es6",
+    "lint:core": "eslint src/Slim.js",
+    "lint:directives": "eslint src/directives",
+    "lint": "npm run lint:core && npm run lint:directives",
+    "build": "npm run lint && node build.js",
     "build:tests": "babel ./tests/decorators/decorators.test-app.js --plugins transform-decorators-legacy,transform-class-properties > ./tests/decorators/decorators.test-app.compiled.js",
     "pretest": "npm run build && npm run build:tests && node prepare-test.js",
     "test": "nightwatch -- tests/*.test.js",
@@ -57,6 +56,12 @@
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-es2015-classes": "^6.24.1",
     "babel-preset-env": "^1.6.1",
+    "eslint": "^5.5.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
     "nightwatch": "0.9.8",
     "selenium-download": "^2.0.7",
     "selenium-server-standalone-jar": "^3.0.1",
