+++ tests/decorators/decorators.test-app.compiled.js
@@ -1,0 +1,78 @@
+'use strict';
+
+var _dec, _dec2, _dec3, _class, _desc, _value, _class2, _descriptor;
+
+function _initDefineProp(target, property, descriptor, context) {
+    if (!descriptor) return;
+    Object.defineProperty(target, property, {
+        enumerable: descriptor.enumerable,
+        configurable: descriptor.configurable,
+        writable: descriptor.writable,
+        value: descriptor.initializer ? descriptor.initializer.call(context) : void 0
+    });
+}
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+function _applyDecoratedDescriptor(target, property, decorators, descriptor, context) {
+    var desc = {};
+    Object['ke' + 'ys'](descriptor).forEach(function (key) {
+        desc[key] = descriptor[key];
+    });
+    desc.enumerable = !!desc.enumerable;
+    desc.configurable = !!desc.configurable;
+
+    if ('value' in desc || desc.initializer) {
+        desc.writable = true;
+    }
+
+    desc = decorators.slice().reverse().reduce(function (desc, decorator) {
+        return decorator(target, property, desc) || desc;
+    }, desc);
+
+    if (context && desc.initializer !== void 0) {
+        desc.value = desc.initializer ? desc.initializer.call(context) : void 0;
+        desc.initializer = undefined;
+    }
+
+    if (desc.initializer === void 0) {
+        Object['define' + 'Property'](target, property, desc);
+        desc = null;
+    }
+
+    return desc;
+}
+
+function _initializerWarningHelper(descriptor, context) {
+    throw new Error('Decorating class property failed. Please ensure that transform-class-properties is enabled.');
+}
+
+var TestApp = (_dec = tag('test-app'), _dec2 = template('\n<h1 bind>{{myProp}}</h1>\n'), _dec3 = attribute(), _dec(_class = _dec2(_class = (_class2 = function (_Slim) {
+    _inherits(TestApp, _Slim);
+
+    function TestApp() {
+        var _ref;
+
+        var _temp, _this, _ret;
+
+        _classCallCheck(this, TestApp);
+
+        for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+            args[_key] = arguments[_key];
+        }
+
+        return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = TestApp.__proto__ || Object.getPrototypeOf(TestApp)).call.apply(_ref, [this].concat(args))), _this), _initDefineProp(_this, 'myProp', _descriptor, _this), _temp), _possibleConstructorReturn(_this, _ret);
+    }
+
+    return TestApp;
+}(Slim), (_descriptor = _applyDecoratedDescriptor(_class2.prototype, 'myProp', [_dec3], {
+    enumerable: true,
+    initializer: function initializer() {
+        return '';
+    }
+})), _class2)) || _class) || _class);
+
